Feature to add new group to group field instantly
